requirements.txt
beautifulsoup4==4.8.0
Flask==1.1.1
grpcio==1.24.1
Flask-Cors==3.0.8
python-dateutil==2.8.0
python-docx==0.8.10
regex==2019.8.19
requests==2.22.0
tika==1.22
wikipedia~=1.4.0
